refactor(events): alert events in separate file
